[TASK] split up of utils, sqr as templated function
